fix: fix ts problem

+++ nuxt/server/api/articles/[slug].delete.ts
@@ -1,14 +1,16 @@
 import { Article } from '~/server/models/Article.model';
 
 export default defineEventHandler(async (event) => {
-	const slug = event.context.params.slug;
-	// delete article
-	try {
-		await Article.findOneAndDelete({ slug: slug });
-		return { message: 'Article deleted' };
-	} catch (e: any) {
-		throw createError({
-			message: e.message,
-		});
+	if (event.context && event.context.params) {
+		const slug = event.context.params.slug;
+		// delete article
+		try {
+			await Article.findOneAndDelete({ slug: slug });
+			return { message: 'Article deleted' };
+		} catch (e: any) {
+			throw createError({
+				message: e.message,
+			});
+		}
 	}
 });
